libva: enable dummy psb video for debug purpose
BZ: 74575

When debug some video bugs, for example repeat camera recoding,
video decode, encode, camera, display, 3d are all run together.
So it is difficult to identify the mistaken module.
For such kind of problem, dummy driver can be used to bypass video
decoding.
For the usage of dummy, please refer to README.

Change-Id: I019fded267963def9f6dfffaf0548709efb94886
Signed-off-by: Fei Jiang <fei.jiang@intel.com>
Signed-off-by: pingshix <pingx.shi@intel.com>
Signed-off-by: Fei Jiang <fei.jiang@intel.com>
Reviewed-on: http://android.intel.com:8080/83961
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ma, Hongpeng <hongpeng.ma@intel.com>
Reviewed-by: Feng, Wei <wei.feng@intel.com>
Reviewed-by: Guo, Nana N <nana.n.guo@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>